Separate EventArgs implementations and EventHandlers
