+++ src/Buddies.App/Components/Carousel.razor
@@ -1,4 +1,5 @@
-﻿@using Microsoft.AspNetCore.Components.Rendering
+﻿@using Buddies.App.Events
+@using Microsoft.AspNetCore.Components.Rendering
 
 @implements IDisposable
