+++ src/Buddies.App/Events/CarouselPageScrollEventArgs.cs
@@ -1,0 +1,6 @@
+﻿namespace Buddies.App.Events;
+
+public class CarouselPageScrollEventArgs : EventArgs
+{
+
+}
