+++ src/Buddies.App/Events/EventHandlers.cs
@@ -1,15 +1,9 @@
 ﻿using Microsoft.AspNetCore.Components;
 
-namespace Buddies.App;
+namespace Buddies.App.Events;
 
 [EventHandler("oncarouselpagescroll", typeof(CarouselPageScrollEventArgs), enableStopPropagation: true, enablePreventDefault: true)]
 public static class EventHandlers
 {
 
 }
-
-public class CarouselPageScrollEventArgs : EventArgs
-{
-
-}
-
